WARNING - Some physics related changes: - Various types of variables and struct members were changed to int16_t or uint16_t, depending on what was (probably) originally used. - In particular, the common type of ck_oldrects and CK_object's clipRects, which is still named CK_objPhysData, is now a struct type with uint16_t fields. On the other hand,  ck_deltarects was changed to have a new struct type of CK_objPhysDataDelta, with just unit position fields of type int16_t. - For the latter, the tile fields were unused, and apparently don't even exist in the original codebase.
At least these don't seem to change much (if any) during the playback
of the demos. These changes can assist with reproducing some variant
of a familiar glitch in Keen 4's level 10 (although not necessarily
in the exact same way as the original EXE does).

+++ src/id_ca.c
@@ -1001,17 +1001,17 @@ void CA_LoadAllSounds(void)
 //extern int ck_currentMapNumber;
 //UPDATE (Mar 6 2014): The var above is now ck_gameState.currentLevel.
 
-uint16_t *CA_TilePtrAtPos(int x, int y, int plane)
+uint16_t *CA_TilePtrAtPos(int16_t x, int16_t y, int16_t plane)
 {
 	return &CA_mapPlanes[plane][y*CA_MapHeaders[ca_mapOn]->width+x];
 }
 
-uint16_t CA_TileAtPos(int x, int y, int plane)
+uint16_t CA_TileAtPos(int16_t x, int16_t y, int16_t plane)
 {
 	return CA_mapPlanes[plane][y*CA_MapHeaders[ca_mapOn]->width+x];
 }
 
-void CA_SetTileAtPos(int x, int y, int plane, int value)
+void CA_SetTileAtPos(int16_t x, int16_t y, int16_t plane, uint16_t value)
 {
 	CA_mapPlanes[plane][y*CA_MapHeaders[ca_mapOn]->width+x] = value;
 }
